requirements.txt
keyboard==0.13.5
MouseInfo==0.1.3
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pillow==10.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
pytweening==1.2.0
